+++ src/hooks/menu_layer.cpp
@@ -27,7 +27,7 @@ bool HookedMenuLayer::init() {
     // reset integrity check state here
     globed::resetIntegrityCheck();
 
-    if (!globed::softDisabled()) {
+    if (!globed::hasSeverelyBrokenResources()) {
         // auto connect
         util::misc::callOnce("menu-layer-init-autoconnect", []{
             auto& settings = GlobedSettings::get();
@@ -163,7 +163,7 @@ void HookedMenuLayer::maybeUpdateButton(float) {
 }
 
 void HookedMenuLayer::onGlobedButton(CCObject*) {
-    if (globed::softDisabled()) {
+    if (globed::hasSeverelyBrokenResources()) {
         geode::createQuickPopup(
             "Globed Error",
             "<cy>Outdated resources</c> were detected. The mod has been <cr>disabled</c> to prevent crashes.\n\nIf you have any <cg>texture packs</c>, or mods that change textures, please try <cy>disabling</c> them.",